Refactor the handling of nested types. (#7)
While building this with clang static analysis, it was
complaining that the 'ret' value in the nested type
handling was being ignored, which is correct.  But while
reviewing the code, I realized that it would be much easier
to read, and hence to review for these kinds of violations,
if this was split into a separate function.  So I've gone
ahead and done that here, and then fixed the bug that
caused us to ignore the return value.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>