uchar: fix conditional include/typedef (#10)
Fix incorrect check: if `__has_include(..)` was supported, but `uchar.h` did not exist, `char16_t` never got `typedef`ed.

It needs to be `typedef`ed in (at least) two cases:

1. `uchar.h` does not exist, and
2. `__has_include(..)` is not supported

In case of the former, the type is most likely missing. In case of the latter, `uchar.h` could exist, but existence can't be checked.

Note: checking for `__has_include(..)` support and use of the macro in a combined conditional is non-portable according to the documentation, hence the split.